data_preprocessing/requirements.txt
google-cloud-storage==2.0.0
opencv-python-headless==4.5.5.62
numpy==1.23.5
tqdm